Files_Versioning-MinIO/requirements.txt
certifi==2022.12.7
config==0.5.1
grpcio==1.51.3
grpcio-tools==1.51.3
minio==7.1.13
protobuf==4.22.0
urllib3==2.0.5
